Updates to examples

+++ examples/Animation_bareMinimum/MyAnimation.cpp
@@ -12,7 +12,7 @@ void MyAnimation::draw(PixelStrip *strip) {
     strip->setPixelColor(0, (pixelState%3==0) ? DARKGREEN : BLACK);
     pixelState = (pixelState + 1) % 3;
     for (int pixelNum = 0; pixelNum < strip->numPixels(); pixelNum++) {
-      unsigned long pixelColor = ((pixelNum + pixelState) % 3 == 0) ? color : BLACK;
+      uint32_t pixelColor = ((pixelNum + pixelState) % 3 == 0) ? color : BLACK;
       strip->setPixelColor(pixelNum, pixelColor);
     }
     setTimeout(waitTime);
